refactor: move vfn service into module
